(feat,wip) move towards cosine sim with expected intensities
