+++ Cargo.toml
@@ -5,8 +5,9 @@ edition = "2021"
 license = "Apache-2.0"
 
 [dependencies]
-timsrust = "0.4.1"
+timsrust = "=0.4.1"
 rayon = "1.5"
+nohash-hasher = "=0.2.0"
 indicatif = { version = "0.17.8", features = ["rayon"] }
 
 clap = { version = "4.5.17", features = ["derive"], optional = true }
