+++ src/models/aggregators/raw_peak_agg/mod.rs
@@ -5,7 +5,6 @@ pub mod point_agg;
 pub use chromatogram_agg::ChromatomobilogramStats;
 pub use multi_chromatogram_agg::MultiCMGStatsAgg;
 // TODO: reorganize this so I donr use direcly from `base``
-pub use multi_chromatogram_agg::base::PartitionedCMGArrays;
 pub use multi_chromatogram_agg::MultiCMGStatsFactory;
 pub use point_agg::{
     RawPeakIntensityAggregator,
