+++ src/models/aggregators/raw_peak_agg/multi_chromatogram_agg/mod.rs
@@ -1,6 +1,8 @@
-pub mod base;
+pub mod aggregator;
+pub mod arrays;
 #[allow(clippy::module_inception)]
 pub mod multi_chromatogram_agg;
+pub mod scored_arrays;
 
 pub use multi_chromatogram_agg::{
     MultiCMGStatsAgg,
