+++ src/models/aggregators/raw_peak_agg/multi_chromatogram_agg/multi_chromatogram_agg.rs
@@ -1,9 +1,11 @@
 use super::super::chromatogram_agg::ScanTofStatsCalculatorPair;
-use super::base::{
-    ParitionedCMGAggregator,
+use super::aggregator::ParitionedCMGAggregator;
+use super::arrays::{
     PartitionedCMGArrayStats,
-    PartitionedCMGScoredStatsArrays,
+    PartitionedCMGArrays,
 };
+use super::scored_arrays::PartitionedCMGScoredStatsArrays;
+use crate::models::elution_group::ElutionGroup;
 use crate::models::frames::raw_peak::RawPeak;
 use crate::models::queries::MsLevelContext;
 use crate::traits::aggregator::Aggregator;
@@ -16,30 +18,53 @@ use timsrust::converters::{
 };
 
 #[derive(Debug, Clone)]
-pub struct MultiCMGStatsAgg<FH: Clone + Eq + Serialize + Hash + Send + Sync> {
+pub struct MultiCMGStatsAgg<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> {
     pub converters: (Tof2MzConverter, Scan2ImConverter),
     pub ms1_stats: ParitionedCMGAggregator<usize>,
     pub ms2_stats: ParitionedCMGAggregator<FH>,
     pub id: u64,
     pub context: Option<MsLevelContext<usize, FH>>,
     pub buffer: Option<ScanTofStatsCalculatorPair>,
+    // TODO: Make this a reference instead of a clone.... maybe.
+    pub elution_group_ref: ElutionGroup<FH>,
 }
 
 #[derive(Debug, Clone)]
-pub struct MultiCMGStatsFactory<FH: Clone + Eq + Serialize + Hash + Send + Sync> {
+pub struct MultiCMGStatsFactory<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> {
     pub converters: (Tof2MzConverter, Scan2ImConverter),
     pub _phantom: std::marker::PhantomData<FH>,
 }
 
-impl<FH: Clone + Eq + Serialize + Hash + Send + Sync> MultiCMGStatsFactory<FH> {
-    pub fn build(&self, id: u64) -> MultiCMGStatsAgg<FH> {
+impl<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> MultiCMGStatsFactory<FH> {
+    pub fn build_with_elution_group(
+        &self,
+        elution_group: &ElutionGroup<FH>,
+    ) -> MultiCMGStatsAgg<FH> {
+        // TODO: RN this is a super hacky ... IDEALLY we would either keep a reference OR
+        // preserve only the expected intensities.
+        let elution_group_ref = elution_group.clone();
+
+        let fragment_keys = elution_group
+            .fragment_mzs
+            .keys()
+            .cloned()
+            .collect::<Vec<FH>>();
+
+        let precursor_keys: Vec<usize> = elution_group
+            .precursor_mzs
+            .iter()
+            .enumerate()
+            .map(|x| x.0)
+            .collect();
+
         MultiCMGStatsAgg {
             converters: (self.converters.0, self.converters.1),
-            ms1_stats: Default::default(),
-            ms2_stats: Default::default(),
-            id,
+            ms1_stats: ParitionedCMGAggregator::new(precursor_keys),
+            ms2_stats: ParitionedCMGAggregator::new(fragment_keys),
+            id: elution_group.id,
             context: None,
             buffer: None,
+            elution_group_ref,
         }
     }
 }
@@ -71,33 +96,21 @@ impl<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> Aggregat
         let rt_miliseconds = (peak.retention_time * 1000.0) as u32;
 
         match self.get_context() {
-            MsLevelContext::MS1(i) => {
-                self.ms1_stats.uniq_rts.insert(rt_miliseconds);
-                self.ms1_stats
-                    .scan_tof_mapping
-                    .entry((i, rt_miliseconds))
-                    .and_modify(|curr| {
-                        curr.add(u64_intensity, peak.scan_index, peak.tof_index);
-                    })
-                    .or_insert(ScanTofStatsCalculatorPair::new(
-                        u64_intensity,
-                        peak.scan_index,
-                        peak.tof_index,
-                    ));
+            MsLevelContext::MS1(_i) => {
+                self.ms1_stats.add(
+                    rt_miliseconds,
+                    peak.scan_index,
+                    peak.tof_index,
+                    u64_intensity,
+                );
             }
-            MsLevelContext::MS2(i) => {
-                self.ms2_stats.uniq_rts.insert(rt_miliseconds);
-                self.ms2_stats
-                    .scan_tof_mapping
-                    .entry((i, rt_miliseconds))
-                    .and_modify(|curr| {
-                        curr.add(u64_intensity, peak.scan_index, peak.tof_index);
-                    })
-                    .or_insert(ScanTofStatsCalculatorPair::new(
-                        u64_intensity,
-                        peak.scan_index,
-                        peak.tof_index,
-                    ));
+            MsLevelContext::MS2(_i) => {
+                self.ms2_stats.add(
+                    rt_miliseconds,
+                    peak.scan_index,
+                    peak.tof_index,
+                    u64_intensity,
+                );
             }
         }
     }
@@ -105,10 +118,10 @@ impl<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> Aggregat
     fn set_context(&mut self, context: Self::Context) {
         match &context {
             MsLevelContext::MS1(i) => {
-                self.ms1_stats.uniq_ids.insert(*i);
+                self.ms1_stats.set_context(i.clone()).unwrap();
             }
             MsLevelContext::MS2(i) => {
-                self.ms2_stats.uniq_ids.insert(i.clone());
+                self.ms2_stats.set_context(i.clone()).unwrap();
             }
         }
         self.context = Some(context);
@@ -130,12 +143,12 @@ impl<FH: Clone + Eq + Serialize + Hash + Send + Sync + std::fmt::Debug> Aggregat
         let mobility_converter = &self.converters.1;
 
         let ms1_stats = PartitionedCMGScoredStatsArrays::new(
-            PartitionedCMGArrayStats::from(self.ms1_stats.finalize()),
+            PartitionedCMGArrayStats::from(PartitionedCMGArrays::from(self.ms1_stats)),
             mz_converter,
             mobility_converter,
         );
         let ms2_stats = PartitionedCMGScoredStatsArrays::new(
-            PartitionedCMGArrayStats::from(self.ms2_stats.finalize()),
+            PartitionedCMGArrayStats::from(PartitionedCMGArrays::from(self.ms2_stats)),
             mz_converter,
             mobility_converter,
         );
