+++ src/models/indices/transposed_quad_index/quad_splitted_transposed_index.rs
@@ -468,7 +468,7 @@ impl<FH: Eq + Hash + Copy + Serialize + Send + Sync>
     }
 }
 
-impl<FH: Copy + Clone + Serialize + Eq + Hash + Send + Sync>
+impl<FH: Copy + Clone + Serialize + Eq + Hash + Send + Sync + std::fmt::Debug>
     ToleranceAdapter<FragmentGroupIndexQuery<FH>, ElutionGroup<FH>>
     for QuadSplittedTransposedIndex
 {
