+++ src/traits/tolerance.rs
@@ -26,7 +26,7 @@ pub enum MobilityTolerance {
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum QuadTolerance {
-    Absolute((f32, f32, u8)),
+    Absolute((f32, f32)),
 }
 
 // TODO: Rename to something that does not use the 'Default'
@@ -36,8 +36,6 @@ pub struct DefaultTolerance {
     pub rt: RtTolerance,
     pub mobility: MobilityTolerance,
     pub quad: QuadTolerance,
-    pub num_ms1_isotopes: usize,
-    pub num_ms2_isotopes: usize,
 }
 
 impl Default for DefaultTolerance {
@@ -46,50 +44,16 @@ impl Default for DefaultTolerance {
             ms: MzToleramce::Ppm((20.0, 20.0)),
             rt: RtTolerance::Absolute((5.0, 5.0)),
             mobility: MobilityTolerance::Pct((3.0, 3.0)),
-            quad: QuadTolerance::Absolute((0.1, 0.1, 1)),
-            num_ms1_isotopes: 3,
-            num_ms2_isotopes: 1,
+            quad: QuadTolerance::Absolute((0.1, 0.1)),
         }
     }
-}
-
-const PROTON_MASS: f64 = 1.007276;
-const NEUTRON_MASS: f64 = 1.008664;
-
-fn mass_to_isotope_mzs(mass: f64, charge: u8, num_isotopes: usize) -> Vec<f64> {
-    let mut out = Vec::with_capacity(num_isotopes);
-    let real_mass = mass + (charge as f64 * PROTON_MASS);
-    for i in 0..(num_isotopes) {
-        let mass = real_mass + (i as f64 * NEUTRON_MASS);
-        let mz = mass / charge as f64;
-        out.push(mz);
-    }
-    out
-}
-
-fn mz_to_isotope_mzs(mz: f64, charge: u8, num_isotopes: usize) -> Vec<f64> {
-    let mut out = Vec::with_capacity(num_isotopes);
-    let proton_mass_frac = PROTON_MASS / charge as f64;
-    for i in 0..num_isotopes {
-        let mz = mz + (i as f64 * proton_mass_frac);
-        out.push(mz);
-    }
-    out
 }
 
 pub trait Tolerance {
     fn mz_range(&self, mz: f64) -> (f64, f64);
     fn rt_range(&self, rt: f32) -> Option<(f32, f32)>;
     fn mobility_range(&self, mobility: f32) -> Option<(f32, f32)>;
-    fn quad_range(&self, precursor_mz: f64, precursor_charge: u8) -> (f32, f32);
-    fn num_ms1_isotopes(&self) -> usize;
-    fn isotope_mzs_mass(&self, monoisotopic_mass: f64, charge: u8) -> Vec<f64> {
-        mass_to_isotope_mzs(monoisotopic_mass, charge, self.num_ms1_isotopes())
-    }
-    fn isotope_mzs_mz(&self, mz: f64, charge: u8) -> Vec<f64> {
-        mz_to_isotope_mzs(mz, charge, self.num_ms1_isotopes())
-    }
-    fn num_ms2_isotopes(&self) -> usize;
+    fn quad_range(&self, precursor_mz_range: (f64, f64)) -> (f32, f32);
 }
 
 impl Tolerance for DefaultTolerance {
@@ -129,39 +93,21 @@ impl Tolerance for DefaultTolerance {
         }
     }
 
-    fn quad_range(&self, precursor_mz: f64, precursor_charge: u8) -> (f32, f32) {
-        // Should this be a recoverable error?
-        if precursor_charge == 0 {
-            panic!(
-                "Precursor charge is 0, inputs: self: {:?}, precursor_mz: {:?}, precursor_charge: {:?}",
-                self, precursor_mz, precursor_charge
-            );
-        };
+    fn quad_range(&self, precursor_mz_range: (f64, f64)) -> (f32, f32) {
         match self.quad {
-            QuadTolerance::Absolute((low, high, num_isotopes)) => {
-                let max_offset = (1.0 / precursor_charge as f32) * num_isotopes as f32;
-                let f32_mz = precursor_mz as f32;
-                let mz_low = f32_mz - low - max_offset;
-                let mz_high = f32_mz + high + max_offset;
+            QuadTolerance::Absolute((low, high)) => {
+                let mz_low = precursor_mz_range.0.min(precursor_mz_range.1) as f32 - low;
+                let mz_high = precursor_mz_range.1.max(precursor_mz_range.0) as f32 + high;
                 assert!(mz_low <= mz_high);
                 assert!(
                     mz_low > 0.0,
-                    "Precursor mz is 0 or less, inputs: self: {:?}, precursor_mz: {:?}, precursor_charge: {:?}",
+                    "Precursor mz is 0 or less, inputs: self: {:?}, precursor_mz_range: {:?}",
                     self,
-                    precursor_mz,
-                    precursor_charge
+                    precursor_mz_range,
                 );
                 (mz_low, mz_high)
             }
         }
-    }
-
-    fn num_ms1_isotopes(&self) -> usize {
-        self.num_ms1_isotopes
-    }
-
-    fn num_ms2_isotopes(&self) -> usize {
-        self.num_ms2_isotopes
     }
 }
 
@@ -172,69 +118,4 @@ impl Tolerance for DefaultTolerance {
 /// an `ElutionGroup` struct.
 pub trait ToleranceAdapter<QF, T> {
     fn query_from_elution_group(&self, tol: &dyn Tolerance, elution_group: &T) -> QF;
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_isotope_mzs_neutral() {
-        let test_vals = vec![
-            (100.0, 1, vec![101.0, 102.0, 103.0]),
-            // For z2, mass would be 102 (100 + 2 protons).... but since charge is 2 then we divide that
-            (100.0, 2, vec![51.0, 51.5, 52.0]),
-            (100.0, 3, vec![34.34, 34.67, 35.0]),
-        ];
-
-        for (monoisotopic_mass, charge, expected) in test_vals {
-            let out = mass_to_isotope_mzs(monoisotopic_mass, charge, 3);
-            assert_eq!(out.len(), expected.len());
-            let abs_diff: Vec<f64> = out
-                .iter()
-                .zip(expected.iter())
-                .map(|(a, b)| (a - b).abs())
-                .collect();
-            for ad in abs_diff.iter() {
-                // Very tight tolerances here ...
-                assert!(
-                    *ad < 0.03,
-                    "Expected {:?}, got {:?}, diff {:?}",
-                    expected,
-                    out,
-                    ad
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn test_isotope_mzs_mz() {
-        let test_vals = vec![
-            (100.0, 1, vec![100.0, 101.0, 102.0]),
-            (100.0, 2, vec![100.0, 100.5, 101.0]),
-            (100.0, 3, vec![100.0, 100.3333, 100.66666]),
-        ];
-
-        for (monoisotopic_mass, charge, expected) in test_vals {
-            let out = mz_to_isotope_mzs(monoisotopic_mass, charge, 3);
-            assert_eq!(out.len(), expected.len());
-            let abs_diff: Vec<f64> = out
-                .iter()
-                .zip(expected.iter())
-                .map(|(a, b)| (a - b).abs())
-                .collect();
-
-            for ad in abs_diff.iter() {
-                // Very tight tolerances here ...
-                assert!(
-                    *ad < 0.03,
-                    "Expected {:?}, got {:?}, diff {:?}",
-                    expected,
-                    out,
-                    ad
-                );
-            }
-        }
-    }
 }
